Fix v4l2loopback detection
This Flatpak does not include modinfo and probably it is not necessary
to include it just to detect if the module is there.

This patch also removes the reference to the v4lsink extension that as
of 26.1 is no longer necessary

The issue of inserting the module still needs to be handled by the user
separately. Until ostree has some support for dkms, the best is probably
that the user builds it and has a systemd service that loads it on the
multiuser target.